Extract GenericPageBreadcrumbsMixin for combining generic views with page views
