+++ wagtail/admin/views/pages/usage.py
@@ -4,7 +4,6 @@ from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import PermissionDenied
 from django.http import Http404
 from django.urls import reverse
-from django.utils.functional import cached_property
 from django.utils.translation import gettext_lazy as _
 
 from wagtail.admin.ui.tables import Column, DateColumn
@@ -16,7 +15,9 @@ from wagtail.admin.ui.tables.pages import (
 )
 from wagtail.admin.views import generic
 from wagtail.admin.views.generic.base import BaseListingView
-from wagtail.admin.views.pages.utils import get_breadcrumbs_items_for_page
+from wagtail.admin.views.pages.utils import (
+    GenericPageBreadcrumbsMixin,
+)
 from wagtail.models import Page
 
 
@@ -73,27 +74,14 @@ class ContentTypeUseView(BaseListingView):
         return context
 
 
-class UsageView(generic.UsageView):
+class UsageView(GenericPageBreadcrumbsMixin, generic.UsageView):
     model = Page
     pk_url_kwarg = "page_id"
     header_icon = "doc-empty-inverse"
     usage_url_name = "wagtailadmin_pages:usage"
     edit_url_name = "wagtailadmin_pages:edit"
-    _show_breadcrumbs = True
 
     def dispatch(self, request, *args, **kwargs):
         if not self.object.permissions_for_user(request.user).can_edit():
             raise PermissionDenied
         return super().dispatch(request, *args, **kwargs)
-
-    @cached_property
-    def breadcrumbs_items(self):
-        return get_breadcrumbs_items_for_page(self.object, self.request.user)
-
-    def get_breadcrumbs_items(self):
-        # The generic UsageView will add an edit link for the current object as
-        # the second-to-last item, but we don't want that because we want to
-        # link to the explore view instead for consistency with how page
-        # breadcrumbs work elsewhere. So we only take the last item, which is
-        # the "self" (Usage) item.
-        return self.breadcrumbs_items + [super().get_breadcrumbs_items()[-1]]
